requirement.txt
langchain
langgraph
langchain-google-genai
sentence-transformers
faiss-cpu
python-dotenv
uvicorn